+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Service Account policies [#82](https://github.com/losisin/gcp-pac/pull/82)
+  - `disallowUserManagedKeys`
+
 ### Changed
 
 - Update packages [#81](https://github.com/losisin/gcp-pac/pull/81)